+++ eagerx_dcsc_setups/pendulum/real/engine_states.py
@@ -41,4 +41,3 @@ class DummyState(EngineState):
 
     def reset(self, state, done):
         pass
-