+++ examples/example_ode.py
@@ -30,7 +30,7 @@ if __name__ == "__main__":
         "Pendulum",
         "pendulum",
         render_shape=[480, 480],
-        sensors=["pendulum_output", "action_applied"],
+        sensors=["x"],
         states=["model_state", "model_parameters"],
     )
     # Visualize EngineGraph
@@ -51,9 +51,8 @@ if __name__ == "__main__":
 
     # Connect the nodes
     graph.connect(action="action", target=bf.inputs.signal)
-    graph.connect(source=bf.outputs.filtered, target=pendulum.actuators.pendulum_input)
-    graph.connect(source=pendulum.sensors.pendulum_output, observation="observation", window=1)
-    graph.connect(source=pendulum.sensors.action_applied, observation="action_applied", window=1)
+    graph.connect(source=bf.outputs.filtered, target=pendulum.actuators.u)
+    graph.connect(source=pendulum.sensors.x, observation="observation", window=1)
 
     # Add rendering
     graph.add_component(pendulum.sensors.image)
@@ -68,23 +67,25 @@ if __name__ == "__main__":
     # Define step function
     def step_fn(prev_obs, obs, action, steps):
         state = obs["observation"][0]
+        u = action["action"][0]
+
         # Calculate reward
         sin_th, cos_th, thdot = state
         th = np.arctan2(sin_th, cos_th)
-        cost = th**2 + 0.1 * (thdot / (1 + 10 * abs(th))) ** 2
+        cost = th**2 + 0.1 * (thdot / (1 + 10 * abs(th))) ** 2 + 0.01 * u ** 2
         # Determine done flag
         done = steps > 500
         # Set info:
-        info = dict()
+        info = {"TimeLimit.truncated": done}
         return obs, -cost, done, info
 
     # Initialize Environment
     env = Flatten(
-        EagerxEnv(name="rx", rate=rate, graph=graph, engine=engine, step_fn=step_fn)
+        EagerxEnv(name="ode_env", rate=rate, graph=graph, engine=engine, step_fn=step_fn)
     )
 
     # Initialize learner (kudos to Antonin)
-    model = sb.SAC("MlpPolicy", env, verbose=1, device="cpu")
+    model = sb.SAC("MlpPolicy", env, verbose=1)
 
     # First train in simulation for 5 minutes and save
     env.render("human")